Refactor: separate FPLang.v into FPCore.v + FPLang.v
The purpose is to permit some clients to import only the
shallow-embedded semantic definitions BPLUS, BMULT, etc.
(which are now in FPCore.v)
without also importing the whole VCFloat system including
the deep-embedded definitions PLUS, MULT, etc.
(which remain in FPLang.v).

Why might such clients exist?  We envision a multilayer correctness
proof of a C program: prove in VST that a C program implements a
functional model (a Gallina function over Flocq floating point);
then prove that the functional model satisfies a high-level
specification (approximates real-valued function F to some accuracy
bound).  It is in that second proof that VCFloat might be used.
The first proof (VST refinement) should not need to import VCFloat.
But the functional model would still like to make use of some
convenient definitions for Flocq functions on floating-point
numbers; those definitions are defined here in FPCore.  Thus, the
refinement proof in VST will need to import FPCore but not the rest
of VCFloat.