New feature: Prune.v contains the prune_terms tactic
Generally before calling the interval tactic, one might
want to call "field_simplify".  But this can explode into a multinomial
with a huge number of terms, most of which are tiny (i.e., higher-order terms).
The prune_terms tactic first reifies (into the Interval tactic's
reified term language), then does its own kind of field_simplify
on the reified term (which may blow up a bit, but in the reified
language, not in Coq terms); then soundly deletes all the
(provably) tiny terms and replaces them with one term that's
the upper bound of all of them.

Prune.v is entirely implemented and mostly proved correct in this commit.

Also in this commit, some minor tweaks in Automate.v.