Use "Derive" for all user proofs that compute a bound.
The "Derive ... SuchThat ... As" feature of Coq is just
the right thing for computing an error bound at the same time
one is proving it correct.  The advantage over our previous method
(that used a sigma type) is that the bound can be transparent
while the proof is opaque.  In particular, the proof term is _so_
opaque that it will not be loaded from a .vo file using Coq's standard
"dont-load-proofs" mode.