Ungeneralize!
Start the process of simplifying the way Rounding.v treats "shiftmaps",
that is, maps from epsilon-delta variables to (various types).
Previously it was very much typeclass-parameterized in a way that
interferes with understanding and does not seem useful.
Now we will just specialize to CompCert-style PTrees on positive numbers.
In this commit, removed one parameter to the Map typeclass,
and remove the MapIndex typeclass entirely.  In a future commit,
probably remove the Map typeclass entirely.

All our current test cases continue to work with no changes.
However, this may make VCFloat incompatible with the VCFloat1.0-style
CompCert Clight interface.  That's fine, we haven't been supporting that
for a long time now, and it may be time to remove it.