Adjust statement of relate_fold_add lemma,
used by the prune_terms tactic, to match the statement of
relate_fold_add that appears in my recent pull request to coq-mmaps.
https://github.com/coq-community/coq-mmaps/pull/14

That way, when coq-mmaps is (soon) made part of the coq platform,
we can easily switch vcfloat from using FMaps (the old, obsolete version
of functional maps) to the new MMaps (which is better).